requirements.txt
django~=1.8.0
django-extensions==1.5.7
djangorestframework==3.2.3
dj-database-url==0.4.1
gunicorn==19.3.0
python-dotenv==0.1.3
pyyaml==3.11
hashids==1.1.0
two1==3.9.10
twilio==5.4.0
phonenumbers==7.4.4
